int value age, weight, height, for consistency

+++ UI/Areas/PatientCaseReview/Pages/ViewModels/PatientViewModel.cs
@@ -9,8 +9,8 @@ namespace UI.Areas.PatientCaseReview.Pages.ViewModels
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
         public string Gender { get; set; }
-        public Int16 Age { get; set; }
-        public Int16 Weight { get; set; }
-        public Int16 Height { get; set; }
+        public int Age { get; set; }
+        public int Weight { get; set; }
+        public int Height { get; set; }
     }
 }
